+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Fixed
 
 - Fixed issue preventing some builds from being edited.
+- Fixed issue where I broke relic fragments on the Item Tracker for about the 40th time. (eSeFBeeSuite)
 
 ## 2024-05-06